+++ server/src/main/java/server/haengdong/domain/action/ActionRepository.java
@@ -1,12 +1,11 @@
-package server.haengdong.persistence;
+package server.haengdong.domain.action;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import server.haengdong.domain.Action;
-import server.haengdong.domain.Event;
+import server.haengdong.domain.event.Event;
 
 @Repository
 public interface ActionRepository extends JpaRepository<Action, Long> {
