+++ server/src/main/java/server/haengdong/domain/action/BillActionRepository.java
@@ -1,11 +1,10 @@
-package server.haengdong.persistence;
+package server.haengdong.domain.action;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import server.haengdong.domain.BillAction;
-import server.haengdong.domain.Event;
+import server.haengdong.domain.event.Event;
 
 @Repository
 public interface BillActionRepository extends JpaRepository<BillAction, Long> {
