+++ server/src/main/java/server/haengdong/domain/action/MemberActionRepository.java
@@ -1,12 +1,11 @@
-package server.haengdong.persistence;
+package server.haengdong.domain.action;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import server.haengdong.domain.Event;
-import server.haengdong.domain.MemberAction;
+import server.haengdong.domain.event.Event;
 
 @Repository
 public interface MemberActionRepository extends JpaRepository<MemberAction, Long> {
