+++ server/src/main/java/server/haengdong/domain/action/MemberActionStatus.java
@@ -1,4 +1,4 @@
-package server.haengdong.domain;
+package server.haengdong.domain.action;
 
 import java.util.Arrays;
