+++ server/src/main/java/server/haengdong/domain/action/MemberGroupIdProvider.java
@@ -1,4 +1,4 @@
-package server.haengdong.domain;
+package server.haengdong.domain.action;
 
 import org.springframework.stereotype.Component;
