+++ server/src/main/java/server/haengdong/domain/event/Event.java
@@ -1,0 +1,28 @@
+package server.haengdong.domain.event;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class Event {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    private String token;
+
+    public Event(String name, String token) {
+        this.name = name;
+        this.token = token;
+    }
+}
