+++ server/src/main/java/server/haengdong/domain/event/EventRepository.java
@@ -1,9 +1,8 @@
-package server.haengdong.persistence;
+package server.haengdong.domain.event;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import server.haengdong.domain.Event;
 
 @Repository
 public interface EventRepository extends JpaRepository<Event, Long> {
