+++ server/src/main/java/server/haengdong/domain/event/EventStep.java
@@ -1,9 +1,11 @@
-package server.haengdong.domain;
+package server.haengdong.domain.event;
 
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,18 +13,16 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class Event {
+public class EventStep {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Event event;
+
     private String name;
 
-    private String token;
-
-    public Event(String name, String token) {
-        this.name = name;
-        this.token = token;
-    }
+    private Long sequence;
 }
