+++ server/src/main/java/server/haengdong/domain/event/EventTokenProvider.java
@@ -1,4 +1,4 @@
-package server.haengdong.domain;
+package server.haengdong.domain.event;
 
 import java.util.UUID;
 import org.springframework.stereotype.Component;
