+++ server/src/test/java/server/haengdong/application/BillActionServiceTest.java
@@ -10,10 +10,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import server.haengdong.application.request.BillActionAppRequest;
-import server.haengdong.domain.BillAction;
-import server.haengdong.domain.Event;
-import server.haengdong.persistence.BillActionRepository;
-import server.haengdong.persistence.EventRepository;
+import server.haengdong.domain.action.BillAction;
+import server.haengdong.domain.event.Event;
+import server.haengdong.domain.action.BillActionRepository;
+import server.haengdong.domain.event.EventRepository;
 
 @SpringBootTest
 class BillActionServiceTest {
