+++ server/src/test/java/server/haengdong/domain/ActionTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import server.haengdong.domain.action.Action;
+import server.haengdong.domain.event.Event;
 
 class ActionTest {
